Added KTMS Generalised FWER control procedure. Added hardcoded input and user prompts for the number of bootstrap iterations in multiple comparisons corrections tests.
Added a check to see whether the multiple comparisons procedure is appropriate for the comparison (for example permutation tests are appropriate for comparisons against chance level, but not paired-samples comparisons against the actual decoding results and permutation decoding results). Included user feedback when an inappropropriate choice is made, and automatic reversion to Bonferroni correction.