Fixed bugs in prepare_my_vectors_erp.
First bug was that the temp matrices were being cleared too early (as they are used for determining the number of labels for training and test data).
The temp matrices are now cleared after they have been used to create the labels.

Second bug was that code for creating training and test labels (for non SVR data) were removed in an earlier update, and so the SVM classifiers did not have any labels to use.
Code for creating condition labels has now been put back into prepare_my_vectors_er.